+++ apps/playground-web/package.json
@@ -14,13 +14,17 @@
     "@radix-ui/react-dropdown-menu": "^2.0.6",
     "@radix-ui/react-slot": "^1.0.2",
     "@radix-ui/react-tabs": "^1.0.4",
+    "@shikijs/twoslash": "^1.6.0",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "lucide-react": "^0.379.0",
     "next": "14.2.3",
     "next-themes": "^0.3.0",
+    "prettier": "^3.2.5",
     "react": "^18",
     "react-dom": "^18",
+    "server-only": "^0.0.1",
+    "shiki": "^1.6.0",
     "tailwind-merge": "^2.3.0",
     "thirdweb": "workspace:*"
   },
