+++ apps/playground-web/src/app/connect/pay/page.tsx
@@ -1,14 +1,27 @@
-"use client";
-
-import { THIRDWEB_CLIENT } from "@/lib/client";
-import { useTheme } from "next-themes";
+import { StyledPayEmbed } from "@/components/styled-pay-embed";
+import { Button } from "@/components/ui/button";
+import { getBaseUrl } from "@/lib/getBaseUrl";
+import type { Metadata } from "next";
 import Link from "next/link";
-import { PayEmbed } from "thirdweb/react";
-import { Button } from "../../../components/ui/button";
+
+export const metadata: Metadata = {
+  metadataBase: new URL(getBaseUrl()),
+  title: "Integrate Fiat & Cross-Chain Crypto Payments | thirdweb Pay",
+  description:
+    "The easiest way for users to transact in your app. Onramp users in clicks and generate revenue for each user transaction. Integrate for free.",
+  openGraph: {
+    images: [
+      {
+        url: "/og-image-pay.png",
+        width: 1200,
+        height: 630,
+        alt: "Integrate Fiat & Cross-Chain Crypto Payments | thirdweb Pay",
+      },
+    ],
+  },
+};
 
 export default function Page() {
-  const { theme } = useTheme();
-
   return (
     <main className="flex-1 content-center relative">
       <section className="w-full py-12 md:py-24 lg:py-32 xl:py-48 z-10">
@@ -43,10 +56,8 @@ export default function Page() {
             </div>
             <div className="w-full mx-auto my-auto sm:w-full lg:order-last relative flex flex-col space-y-2">
               <div className="shadow-xl md:w-[400px] md:mx-auto">
-                <PayEmbed
-                  client={THIRDWEB_CLIENT}
-                  // we invert here to make it stand out more
-                  theme={theme === "light" ? "dark" : "light"}
+                <StyledPayEmbed
+                // we invert here to make it stand out more
                 />
               </div>
               <p className="md:text-xl text-center text-muted-foreground">
